% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zp_cut_it.R
\name{cutit}
\alias{cutit}
\title{Cut Continuous Vector to Classification}
\usage{
cutit(x, cut_points, include = "low", labels = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{cut_points}{cuting points value}

\item{include}{The direction of cutoff point. Any left letter of lower or upper}

\item{labels}{logical. False is defaulted. TRUE means set range as factor.}
}
\value{
numeric vector or factor
}
\description{
Cut Continuous Vector to Classification
}
\examples{
cutit(mtcars$disp,c(150,190))
cutit(mtcars$disp,c(150,190),labels = TRUE)
}
