% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/em.R
\name{em}
\alias{em}
\title{Expectation-Maximization estimation of a finite mixture model}
\usage{
em(data, D1, D2, t = 1e-64)
}
\arguments{
\item{data}{A vector of real numbers, the data to model with a finite mixture
model.}

\item{D1,D2}{Probability distributions constituting the finite mixture model.
See Details.}

\item{t}{A numerical scalar indicating the value below which the E-M
algorithm should stop.}
}
\value{
A list with class \code{em} containing the following components:
	\item{lambda}{a numerical vector of length \code{length(data)} containing,
   for each datum, the probability to belong to distribution \code{D1}.}
 \item{param}{the location (mu) and scale (sigma) parameters of the
	  probability distributions \code{D1} and \code{D2}.}
   \code{mu2} and \code{lambda2}.}
	\item{D1,D2}{character scalars containing the names of the probability
   distributions used in the finite mixture model.}
	\item{data}{the numerical vector of data used as input.}
	\item{data_name}{character scalar containing the name of the dataset used as
   input.}
	\item{out}{an object of class \code{mle2} that contains the
   maximum-likelihood estimates of parameters \code{mu1}, \code{lambda1},
	\item{t}{the input \code{t} argument value.}
}
\description{
\code{em} returns points estimations of the parameters of a finite mixture
		model using the Expectation-Maximization (E-M) algorithm.
}
\details{
The finite mixture model considered in this function is a mixture of two
 probability distributions that are one of the following: normal, log-normal,
 gamma or Weibull. Each of these distributions is defined by two parameters:
 a location and a scale parameter:
\tabular{lcc}{
             \tab location \tab scale \cr
  normal     \tab mean     \tab sd    \cr
  log-normal \tab meanlog  \tab sdlog \cr
  gamma      \tab shape    \tab rate  \cr
  Weibull    \tab shape    \tab scale
}
These parameters, together with the mixture parameter, are estimated by the
 Expection-Maximization algorithm.
}
\examples{
# Measles IgG concentration data:
length(measles)
range(measles)
# Plotting the data:
hist(measles,100,F,xlab="concentration",ylab="density",ylim=c(0,.55),
  main=NULL,col="grey")
# The kernel density:
lines(density(measles),lwd=1.5,col="blue")
# Estimating the parameters of the finite mixture model:
(measles_out <- em(measles,"normal","normal"))
# The confidence interval of the parameter estimates:
confint(measles_out,t=1e-64,nb=100,level=.95)
# Adding the E-M estimated finite mixture model:
lines(measles_out,lwd=1.5,col="red")
# The legend:
legend("topleft",leg=c("non-parametric","E-M"),col=c("blue","red"),
  lty=1,lwd=1.5,bty="n")
}
\references{
Chuong B. Do and Serafim Batzoglou (2008) What is the expectation
   maximization algorithm? Nature Biotechnology 26(8): 897-899.\cr
 \cr
 Peter Schlattmann (2009) Medical Applications of Finite Mixture Models.
   Springer-Verlag, Berlin.
}
\seealso{
\code{\link{confint.em}} method for calculating the confidence
   intervals of the parameters and \code{\link{cutoff}} for deriving a
   cut-off value.
}

