% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zp_judge.R
\name{judge_123}
\alias{judge_123}
\title{Whether the Data Is Arranged from Small to Large}
\usage{
judge_123(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
logical
}
\description{
Whether the Data Is Arranged from Small to Large
}
\examples{
judge_123(c(1,2,3,4,5))
judge_123(c(1,3,2))
}
