% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zp_judge.R
\name{judge_321}
\alias{judge_321}
\title{Whether the Data Is Arranged from Large to Small}
\usage{
judge_321(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
logical
}
\description{
Whether the Data Is Arranged from Large to Small
}
\examples{
judge_321(c(5,4,3,2,1))
judge_321(c(3,1,2))
}
