% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logrank.R
\name{logrank}
\alias{logrank}
\title{Significant Cutoff Value for Logrank Analysis}
\usage{
logrank(data, time, y, x, cut.numb, n.per, y.per, p.cut = 0.05,
  strict = TRUE, include = "low", round = 2, adjust = 1)
}
\arguments{
\item{data}{data}

\item{time}{name for time variable}

\item{y}{name for y, must be coded as 1 and 0. The outcome must be 1}

\item{x}{name for x}

\item{cut.numb}{number of cutoff points}

\item{n.per}{the least percentage of the smaller group comprised in all patients}

\item{y.per}{the least percentage of the smaller outcome patients comprised in each group}

\item{p.cut}{cutoff of p value, default is 0.05}

\item{strict}{logical. TRUE means significant differences for each group
combination were considered. FALSE means considering for any combination}

\item{include}{direction of cutoff point. Any left letter of lower or upper}

\item{round}{digital. Default is 2}

\item{adjust}{numeric value, adjust methord for p value. 1, defaulted, represents Bonferroni. 2 represent formula given by Douglas G in 1994}
}
\value{
a dataframe contains cutoff points value, subject numbers in each group,
    dumb variable, beta of regression and p value.
}
\description{
Significant Cutoff Value for Logrank Analysis
}
\examples{
logrank(data=mtcars,
        time = 'disp',y='am', x='wt',
        cut.numb=2,
        n.per=0.25,
        y.per=0.10)

logrank(data=mtcars,
        time = 'disp',y='am', x='wt',
        cut.numb=2,
        n.per=0.25,
        y.per=0.10,
        p.cut=0.05,
        strict=TRUE,
        include='low',
        round=2)
}
