% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{mLL}
\alias{mLL}
\title{Minus log-likelihood of the finite mixture model.}
\usage{
mLL(mu1, sigma1, mu2, sigma2, lambda, data, D1, D2)
}
\description{
Minus log-likelihood of the finite mixture model.
}
\keyword{internal}
