% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.R
\name{plothistogram}
\alias{plothistogram}
\title{Plot Hist}
\usage{
plothistogram(marker, cutoff = NULL, gauss = TRUE, npoints = 1000,
  nbreaks = 100, jpg.width = 1000 * sqrt(2), jpg.height = 1000,
  jpg.pointsize = 30, lwd = 3, jpg.file = NULL)
}
\description{
Plot
}

