% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.R
\name{plotHR}
\alias{plotHR}
\title{Plot HR}
\usage{
plotHR(Y, marker, time, event, cutoff = NULL, thres.p = 0.05,
  conf.level = 95, jpg.width = 1000 * sqrt(2), jpg.height = 1000,
  jpg.pointsize = 30, lwd = 3, jpg.file = NULL)
}
\description{
Plot
}

