% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.R
\name{plotkaplanmeier}
\alias{plotkaplanmeier}
\title{Plot KM}
\usage{
plotkaplanmeier(marker, time, event, cutoff = NULL, surv.test = "sctest",
  conf.level = 95, colors = c("black", "red"), jpg.width = 1000 * sqrt(2),
  jpg.height = 1000, jpg.pointsize = 30, lwd = 3, jpg.file = NULL)
}
\description{
Plot
}

