% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.R
\name{plotROC}
\alias{plotROC}
\title{Plot ROC}
\usage{
plotROC(Y, marker = NULL, outcome = NULL, cutoff = NULL,
  jpg.width = 1000 * sqrt(2), jpg.pointsize = 30, lwd = 4,
  jpg.file = NULL)
}
\description{
Plot
}

