% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.R
\name{plotwaterfall}
\alias{plotwaterfall}
\title{Plot Waterfall}
\usage{
plotwaterfall(marker, outcome, cutoff = NULL, conf.level = 95,
  jpg.width = 1000 * sqrt(2), jpg.height = 1000, jpg.pointsize = 30,
  lwd = 4, jpg.file = NULL)
}
\description{
Plot
}

