% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.R
\name{r.get.cutoff}
\alias{r.get.cutoff}
\title{Get cutoff}
\usage{
r.get.cutoff(type = c("none", "distribution", "outcome_significance",
  "outcome_euclidean", "outcome_manhattan", "outcome_sensitivity",
  "outcome_specificity", "survival_significance", "manual"), biomarker = NULL,
  outcome = NULL, time = NULL, event = NULL, cutoff = NULL,
  threshold = NULL, nmin = 10)
}
\description{
Get cutoff
}

