% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{To Get the Best Cutoff Value for ROC Curve}
\usage{
roc(score, class)
}
\arguments{
\item{score}{continuous value}

\item{class}{bianary value, 0 and 1}
}
\value{
If the auc of a variate is lower than 0.5, we treat it as negative
    classification and return information about the negative prediction.
    Otherwise, The variate will be treated as positive one.
}
\description{
Youden index is used for seeking the best cutoff value for ROC Curve.
}
\examples{
roc(score = mtcars$qsec,class = mtcars$am)
roc(score = mtcars$drat,class = mtcars$am)
}
