% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zp_x_ab.R
\name{x_ab}
\alias{x_ab}
\title{Return x Between a and b}
\usage{
x_ab(x, a, b, include = "l")
}
\arguments{
\item{x}{numeric vector}

\item{a}{one number}

\item{b}{one number}

\item{include}{The direction of a and b. Any left letter of lower or upper}
}
\value{
values of x between a and b
}
\description{
Return x Between a and b
}
\examples{
x_ab(mtcars$disp,150,190)
}
