% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Cutoff Value for Logistic Regression}
\usage{
logit(data, y, x, cut.numb, n.per, y.per, p.cut = 0.05, strict = TRUE,
  include = "low", round = 2)
}
\arguments{
\item{data}{data}

\item{y}{name for y, must code as 1 and 0. The outcome must be 1}

\item{x}{name for x}

\item{cut.numb}{number of cutoff points}

\item{n.per}{the least percentage of the smaller group comprised in all patients}

\item{y.per}{the least percentage of the outcome patients comprised in smaller group}

\item{p.cut}{cutoff of p value, default is 0.05}

\item{strict}{logical. TRUE means significant differences for each group
combination were considered. FALSE means considering for any group}

\item{include}{direction of cutoff point. Any left letters of lower or upper}

\item{round}{digital. Default is 2}
}
\value{
a dataframe contains cutoff points value, subject numbers in each group,
    dump variable, or of regression and p value.
}
\description{
Cutoff Value for Logistic Regression
}
\examples{
logit(data=mtcars,
      y='am',
      x='disp',
      cut.numb=1,
      n.per=0.25,
      y.per=0.25)
 logit(data=mtcars,
       y='am',
       x='disp',
       cut.numb=1,
       n.per=0.25,
       y.per=0.20,
       p.cut=0.05,
       strict=TRUE,
       include='low',
       round=2)
}
